Ref: create src folder for sentry_flutter and export files
