+++ flutter/lib/src/default_integrations.dart
