+++ flutter/lib/src/sentry_flutter_web.dart
