+++ flutter/test/default_integrations_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sentry/sentry.dart';
-import 'package:sentry_flutter/default_integrations.dart';
+import 'package:sentry_flutter/src/default_integrations.dart';
 
 import 'mocks.dart';
