+++ flutter/test/file_system_transport_test.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry/sentry.dart';
-import 'package:sentry_flutter/file_system_transport.dart';
+import 'package:sentry_flutter/src/file_system_transport.dart';
 
 void main() {
   const MethodChannel _channel = MethodChannel('sentry_flutter');
